Back to floats

+++ python_introduction/finance_calculator.py
@@ -1,8 +1,8 @@
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expense = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+monthly_expense = float(input("Enter your total monthly expenses: "))
 
 monthly_savings = monthly_income - monthly_expense
-projected_annual_savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
+projected_annual_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
 
 print(f"Your monthly savings are ${monthly_savings}")
 print(f"Projected savings after one year, with interest is: ${projected_annual_savings}.")
